src/main/java/br/com/ajvideira/algafood/api/controller/EstadoController.java
package br.com.ajvideira.algafood.api.controller;

import java.util.List;

import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import br.com.ajvideira.algafood.domain.model.Estado;
import br.com.ajvideira.algafood.domain.repository.EstadoRepository;

@RequestMapping("/estados")
@RestController
public class EstadoController {

    private EstadoRepository estadoRepository;

    public EstadoController(EstadoRepository estadoRepository) {
        this.estadoRepository = estadoRepository;
    }

    public List<Estado> getAll() {
        return this.estadoRepository.findAll();
    }

}
